Add search fields to dismissable filters and refactor export page
