--- app/views/Project/Tagging/Export/ExportPreview/styles.css
@@ -1,32 +1,0 @@
-.export-preview {
-    background-color: var(--dui-color-foreground);
-
-    .header {
-        border: var(--dui-width-separator-thin) solid var(--dui-color-accent);
-        background-color: var(--dui-color-background-information);
-        padding: var(--dui-spacing-small) var(--dui-spacing-medium);
-
-        .heading {
-            color: var(--dui-color-accent);
-        }
-    }
-
-    .main-content {
-        display: flex;
-        flex-direction: column;
-        flex-grow: 1;
-        background-color: var(--dui-color-foreground);
-
-        .label {
-            margin: auto;
-            text-align: center;
-            color: var(--dui-color-text-label);
-            font-size: var(--dui-font-size-large);
-        }
-
-        .lead-preview {
-            flex-grow: 1;
-            width: 100%;
-        }
-    }
-}
